Cleans up the script and moves code to modules
