+++ packages/data-collector/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Feldspar Demo</title>
+    <style>body { padding: 2rem; }</style>
   </head>
   <body>
     <div id="root"></div>
