+++ packages/feldspar/src/framework/visualization/react/ui/elements/spinner.tsx
@@ -29,7 +29,7 @@ export const Spinner = ({
           cy="12"
           r="10"
           stroke="currentColor"
-          stroke-width="4"
+          strokeWidth="4"
         ></circle>
         <path
           className="opacity-75"
