+++ packages/python/port/file_operations.py
@@ -1,0 +1,24 @@
+import zipfile
+
+def get_zipfile(filename):
+    try:
+        return zipfile.ZipFile(filename)
+    except zipfile.error:
+        return None
+
+
+def get_files(zipfile_ref):
+    try:
+        return zipfile_ref.namelist()
+    except zipfile.error:
+        return []
+
+
+def extract_file(zipfile_ref, filename):
+    try:
+        info = zipfile_ref.getinfo(filename)
+        return (filename, info.compress_size, info.file_size)
+    except zipfile.error:
+        return None
+
+
