+++ tests/donation.spec.ts
@@ -7,19 +7,19 @@ import * as path from 'path';
 async function setupTestWithFileUpload(page: Page): Promise<void> {
   // Navigate to the local development server
   await page.goto('http://localhost:3000/');
- 
+
   // Wait for the page to load with increased timeout
-  await expect(page.getByRole('heading', { name: 'Data donation flow example' })).toBeVisible({ timeout: 30000 });
-  
+  await expect(page.getByRole('heading', { name: 'Youtube Data donation' })).toBeVisible({ timeout: 30000 });
+
   // Create a temporary file input for file upload
   const fileChooserPromise = page.waitForEvent('filechooser');
   await page.getByText('Choose file').click();
   const fileChooser = await fileChooserPromise;
-  
+
   // Set a test zip file path
   const zipFilePath = path.join(__dirname, 'test.zip');
   await fileChooser.setFiles(zipFilePath);
-  
+
   // Click continue to process the file
   await page.getByText('Continue').click();
 }
@@ -46,9 +46,9 @@ async function submitDataAndGetResult(page: Page): Promise<string | null> {
 
 test('can submit data', async ({ page }) => {
   await setupTestWithFileUpload(page);
-  
+
   const submittedData = await submitDataAndGetResult(page);
-  
+
   // The submitted data should contain the expected file
   expect(submittedData).toEqual(expect.stringContaining("hello_world.txt"));
 });
@@ -65,7 +65,7 @@ test('can remove rows from submission', async ({ page }) => {
   await expect(page.getByText('hello_world.txt')).not.toBeVisible();
 
   const submittedData = await submitDataAndGetResult(page);
-  
+
   // The submitted data should not contain the deleted file
   expect(submittedData).not.toEqual(expect.stringContaining("hello_world.txt"));
   // The submitted data should contain the other table contents
@@ -81,22 +81,22 @@ test('can undo row removal before submission', async ({ page }) => {
 
   // Toggle the adjust checkbox
   await page.getByRole('checkbox').first().click();
-  
+
   // Select all items for deletion
   const table = await page.getByTestId('table-zip_content');
   await table.getByRole('checkbox').first().click();
 
   await page.getByText('Delete selected').first().click();
   await expect(table.getByText('hello_world.txt')).not.toBeVisible();
-  
+
   // Click the undo button
   await page.getByRole('button', { name: 'Undo' }).click();
-  
+
   // Verify the deleted file is visible again
   await expect(table.getByText('hello_world.txt')).toBeVisible();
 
   const submittedData = await submitDataAndGetResult(page);
-  
+
   // The submitted data should contain the previously deleted file
   expect(submittedData).toEqual(expect.stringContaining("hello_world.txt"));
   // The submitted data should also contain the other table contents
@@ -108,7 +108,7 @@ test('can cancel submission', async ({ page }) => {
 
   // Toggle the adjust checkbox
   await page.getByRole('checkbox').first().click();
-  
+
   // Setup the route to capture the submission data
   const result = setupRouteForDataSubmission(page);
   await page.getByText('No', { exact: true }).click();
